Fix up some library stuff (pow function mainly)

+++ lib/include/number.hpp
@@ -25,7 +25,7 @@ class Number {
 
     /* Conversion */
     std::string toString();
-    int toInt();
+    long long unsigned int toInt();
 
     /* Arithmetic. */
